Refactor to own component
